add many button

+++ TankGame/Assets/Script/HUDController.cs
@@ -8,16 +8,15 @@ public class HUDController : MonoBehaviour {
     public Text gas;
     public float _gas;
     public Text health;
-    public float _health;
+    public int _health;
     public Text force;
     public float _force;
 
     private void LateUpdate()
     {
         gas.text = "Gas: " + (int)_gas;
-        health.text = "Health: " + (int)_health;
-
-        force.text = "Force: " + (int)_force; 
+        health.text = "Health: " + _health;
+        force.text = "Force: " + (int)_force;
     }
 
 }
